# tests/contract/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ElementoCola": {
        "additionalProperties": false,
        "description": "Una derivación esperando a que alguien la recoja.",
        "properties": {
          "context_ref": {
            "title": "Context Ref",
            "type": "string"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "creado_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creado En"
          },
          "cuenta_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuenta Id"
          },
          "motivo_codigo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivo Codigo"
          },
          "resumen_asesor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumen Asesor"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "context_ref",
          "conversation_id"
        ],
        "title": "ElementoCola",
        "type": "object"
      },
...
      "EstadoSala": {
        "additionalProperties": false,
        "description": "Estado de la conversación: quién la atiende y qué se ha dicho.",
        "properties": {
          "asesor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asesor"
          },
          "conversation_id": {
            "title": "Conversation Id",
            "type": "string"
          },
          "cuenta_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cuenta Id"
          },
          "derivada": {
            "default": false,
            "title": "Derivada",
            "type": "boolean"
          },
          "modo": {
            "description": "AUTONOMA cuando responde la IA, ASISTIDA con asesor dentro",
            "title": "Modo",
            "type": "string"
          },
          "resumen_asesor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resumen Asesor"
          },
          "turnos": {
            "items": {
              "$ref": "#/components/schemas/TurnoPublicado"
            },
            "title": "Turnos",
            "type": "array"
          }
        },
        "required": [
          "conversation_id",
          "modo"
        ],
        "title": "EstadoSala",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MensajeAsesor": {
        "additionalProperties": false,
        "description": "Lo que el asesor escribe al cliente dentro de la conversación.",
        "properties": {
          "texto": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Texto",
            "type": "string"
          }
        },
        "required": [
          "texto"
        ],
        "title": "MensajeAsesor",
        "type": "object"
      },
...
      "TurnoPublicado": {
        "additionalProperties": false,
        "description": "Un turno tal y como lo ve la consola del asesor o la del cliente.",
        "properties": {
          "autor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Identificador del asesor, solo en los turnos de rol asesor",
            "title": "Autor"
          },
          "rol": {
            "title": "Rol",
            "type": "string"
          },
          "texto": {
            "title": "Texto",
            "type": "string"
          },
          "ts": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ts"
          }
        },
        "required": [
          "rol",
          "texto"
        ],
        "title": "TurnoPublicado",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "JWT HS256 emitido por /dev/token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "email": "concursos-cis@ulima.edu.pe",
      "name": "Equipo recibo-claro"
    },
...
    "license": {
      "name": "MIT"
    },
    "summary": "Explica la variación del recibo con cero invenciones financieras, verificadas en código y comprobables por logs.",
    "title": "recibo-claro · Explicación inteligente de recibos Movistar",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/asesor/cola": {
      "get": {
        "description": "Cola del 104: casos derivados que nadie ha tomado todavía.\n\nSe excluyen las conversaciones que ya tienen un asesor dentro, porque una cola que\nmuestra lo que otro está atendiendo genera dos personas sobre el mismo cliente.",
        "operationId": "cola_v1_asesor_cola_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ElementoCola"
                  },
                  "title": "Response Cola V1 Asesor Cola Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Derivaciones esperando a que un asesor las recoja",
        "tags": [
          "asesor"
        ]
      }
    },
    "/v1/asesor/conversacion/{conversation_id}": {
      "get": {
        "description": "Devuelve el estado y todos los turnos.\n\nEs lo que sondea la consola del asesor. Se eligió sondeo y no un canal permanente\nporque para una demostración el resultado visible es idéntico y la mitad de piezas\npueden fallar. El salto a eventos del servidor no cambia este contrato.",
        "operationId": "estado_v1_asesor_conversacion__conversation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoSala"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Estado de la sala y transcripción completa",
        "tags": [
          "asesor"
        ]
      }
    },
    "/v1/asesor/conversacion/{conversation_id}/mensaje": {
      "post": {
        "description": "Publica un turno de rol ``asesor``.\n\n**Este texto no pasa por el verificador numérico**, y es deliberado: el verificador\ngarantiza que la *máquina* no invente cifras. Una persona responde de sus propias\npalabras. Lo que sí queda es el rastro nominal en la bitácora.\n\nRaises:\n    ErrorApi: ``409`` si quien escribe no es el asesor que ocupa la sala.",
        "operationId": "mensaje_v1_asesor_conversacion__conversation_id__mensaje_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MensajeAsesor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoSala"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "El asesor escribe al cliente dentro de la conversación",
        "tags": [
          "asesor"
        ]
      }
    },
    "/v1/asesor/conversacion/{conversation_id}/salir": {
      "post": {
        "description": "Libera la sala. El asistente vuelve a modo autónomo.",
        "operationId": "salir_v1_asesor_conversacion__conversation_id__salir_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoSala"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "El asesor deja la sala y la IA vuelve a atender",
        "tags": [
          "asesor"
        ]
      }
    },
    "/v1/asesor/conversacion/{conversation_id}/unirse": {
      "post": {
        "description": "Suma al asesor a la sala. A partir de aquí el asistente no responde al cliente.\n\nRaises:\n    ErrorApi: ``404`` si la conversación no existe; ``409`` si ya hay otro asesor.",
        "operationId": "unirse_v1_asesor_conversacion__conversation_id__unirse_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstadoSala"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "El asesor entra a la conversación y la IA pasa a copiloto",
        "tags": [
          "asesor"
        ]
      }
    },
...
  },
...
}